disable row selection for array sub table page
Row selection has been disabled because bulk delete functionality
cannot work for array sub table according to the current implementation.

Bulk delete problem:

Suppose we have an array with values `[a, b, c]`. Suppose, the indexes
to be removed are [1, 2]. The UI asks firebase function to remove index
1. After 1, the UI asks index 2 to be deleted but index 2 doesn't exist.
After deletion of index 1, the array is `[a, c]`.