improvements such as:
1. Use generateId function instead of Math.random()

2. Remove `tablePage` from useEffect in `Table.tsx`

Upon an error in the firebase query this was causing an infinite loop
to increase page number.

3. Change the logic of when the clear and disable button are shown/disabled

The existing logic was preventing certain tasks to be performed.

For example, deleting all filters individually was disabling both apply
and clear all button. As a user I expect to delete all filters and
click on the apply button to save.